Wait for cardano-node sync
Smoke tests were failing because we would query the protocol-parameters
while the node is not yet in sync and that would throw era missmatch
because of the latest changes to hydra-node. We solve this by waiting
for the cardano-node to be in sync immediately since this workflow is
used only when running smoke-tests. We should investigate if this error
we get makes sense for the potential users trying to run their
hydra-node using the cardano-node that is not fully synced.
